+++ nym-vpn-android/core/src/main/java/net/nymtech/vpn/config/CoreVpnConfigRepository.kt
@@ -29,6 +29,7 @@ class CoreVpnConfigRepository(context: Context) {
 					is CoreVpnConfigUpdate.SetNetwork -> acc.copy(network = update.value)
 					is CoreVpnConfigUpdate.SetDebugLog -> acc.copy(debugLog = update.value)
 					is CoreVpnConfigUpdate.SetSentry -> acc.copy(sentry = update.value)
+					is CoreVpnConfigUpdate.SetLewes -> acc.copy(lewes = update.value)
 				}
 			}
 		}